Rework most of the motion code to be much cleaner and report the actual current position to Stellarium

+++ Dobson.h
@@ -15,8 +15,8 @@
 
 
 struct AzAltPosition {
-	long azimuth;
-	long altitude;
+	double azimuth;
+	double altitude;
 };
